Plugin updated: mobile

+++ upload/source/plugin/mobile/install.php
@@ -5,6 +5,7 @@
  *      This is NOT a freeware, use is subject to license terms
  *
  *      $Id: install.php 31700 2012-09-24 03:46:59Z zhangjie $
+ *	Modified by Valery Votintsev, codersclub.org
  */
 
 if(!defined('IN_DISCUZ')) {
@@ -15,7 +16,7 @@ $sql = <<<EOF
 
 DROP TABLE IF EXISTS pre_common_devicetoken;
 CREATE TABLE pre_common_devicetoken (
-  `uid` mediumint(8) unsigned NOT NULL,
+  `uid` int(11) unsigned NOT NULL,
   `token` char(50) NOT NULL,
   PRIMARY KEY (`uid`),
   KEY `token` (`token`)
